Fix GET for /reviews
All it needed was the database to be specified in the call.

Side note: Nathan and I had issues even getting logged in. We both
had "username" instead of the intended "user" field in credentials.json.
That should be in the gitignore but somehow we both copied over that
error.